refactor: split IndexSettingsManager abstraction into abstractions folder
