+++ packages/api-elasticsearch-tasks/src/index.ts
@@ -7,7 +7,7 @@ import {
     ElasticsearchSynchronize,
     Manager
 } from "~/tasks/index.js";
-import { IndexSettingsManagerDI } from "~/settings/IndexSettingsManager.js";
+import { IndexSettingsManager } from "~/settings/IndexSettingsManager.js";
 import type { PluginCollection } from "@webiny/plugins/types.js";
 import type { Context } from "~/types.js";
 
@@ -15,7 +15,7 @@ export const createElasticsearchBackgroundTasks = (): PluginCollection => {
     return [
         createContextPlugin<Context>(context => {
             context.container.register(Manager);
-            context.container.register(IndexSettingsManagerDI);
+            context.container.register(IndexSettingsManager);
             context.container.register(ElasticsearchReindexingTask);
             context.container.register(ElasticsearchEnableIndexingTask);
             context.container.register(ElasticsearchSynchronize);
