+++ packages/api-elasticsearch-tasks/src/settings/abstractions/IndexSettingsManager.ts
@@ -1,0 +1,15 @@
+import { createAbstraction } from "@webiny/feature/api";
+import type { IIndexSettingsValues } from "~/types.js";
+
+export interface IIndexSettingsManager {
+    getSettings(index: string): Promise<IIndexSettingsValues>;
+    setSettings(index: string, settings: IIndexSettingsValues): Promise<void>;
+}
+
+export const IndexSettingsManager = createAbstraction<IIndexSettingsManager>(
+    "ElasticsearchTasks/IndexSettingsManager"
+);
+
+export namespace IndexSettingsManager {
+    export type Interface = IIndexSettingsManager;
+}
